无良设置 hp、mp short为int. 设置地图id为int

+++ trunk/src/l1j/jrwz/server/WarTimeController.java
@@ -114,7 +114,7 @@ public class WarTimeController implements Runnable {
                             }
                             loc = L1CastleLocation.getGetBackLoc(castleId);
                             L1Teleport.teleport(pc, loc[0], loc[1],
-                                    (short) loc[2], 5, true);
+                                    loc[2], 5, true);
                         }
                     }
                 }
