+++ trunk/src/l1j/jrwz/server/model/L1HauntedHouse.java
@@ -113,7 +113,7 @@ public class L1HauntedHouse {
                 L1SkillUse l1skilluse = new L1SkillUse();
                 l1skilluse.handleCommands(pc, CANCELLATION, pc.getId(),
                         pc.getX(), pc.getY(), null, 0, L1SkillUse.TYPE_LOGIN);
-                L1Teleport.teleport(pc, 32624, 32813, (short) 4, 5, true);
+                L1Teleport.teleport(pc, 32624, 32813, 4, 5, true);
             }
         }
         clearMembers();
